Revert "Allow BCCSP config to be set using env var (#1900)" (#2093)
This reverts commit 0a9f7663a3a548385f3e4fa55e147b6dbce055d4.
which pushed the pkcs11 code outside of the pkcs11 build tag.
This prevents cross-compilation of Fabric due to the fact the
miekg/pkcs11 package contains CGO which requires a cross-compiler
for each platform you are cross compiling for.

We need to find a better approach to solving this problem
that doesn't pull the code out of the build tag.
