+++ integration/pkcs11/pkcs11_suite_test.go
@@ -14,8 +14,6 @@ import (
 	"github.com/hyperledger/fabric/integration"
 	"github.com/hyperledger/fabric/integration/nwo"
 	"github.com/hyperledger/fabric/integration/nwo/commands"
-	"github.com/hyperledger/fabric/integration/nwo/fabricconfig"
-	"github.com/miekg/pkcs11"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/onsi/gomega/gbytes"
@@ -34,9 +32,6 @@ func TestPKCS11(t *testing.T) {
 var (
 	buildServer *nwo.BuildServer
 	components  *nwo.Components
-	ctx         *pkcs11.Ctx
-	sess        pkcs11.SessionHandle
-	bccspConfig *fabricconfig.BCCSP
 )
 
 var _ = SynchronizedBeforeSuite(func() []byte {
@@ -46,9 +41,6 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 	components = buildServer.Components()
 	payload, err := json.Marshal(components)
 	Expect(err).NotTo(HaveOccurred())
-
-	setupPKCS11()
-
 	return payload
 }, func(payload []byte) {
 	err := json.Unmarshal(payload, &components)
@@ -58,64 +50,10 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 var _ = SynchronizedAfterSuite(func() {
 }, func() {
 	buildServer.Shutdown()
-	ctx.Destroy()
-	ctx.CloseSession(sess)
 })
 
 func StartPort() int {
 	return integration.PKCS11Port.StartPortForNode()
-}
-
-func setupPKCS11() {
-	lib, pin, label := bpkcs11.FindPKCS11Lib()
-	ctx, sess = setupPKCS11Ctx(lib, label, pin)
-	bccspConfig = &fabricconfig.BCCSP{
-		Default: "PKCS11",
-		PKCS11: &fabricconfig.PKCS11{
-			Security: 256,
-			Hash:     "SHA2",
-			Pin:      pin,
-			Label:    label,
-			Library:  lib,
-		},
-	}
-}
-
-// Creates pkcs11 context and session
-func setupPKCS11Ctx(lib, label, pin string) (*pkcs11.Ctx, pkcs11.SessionHandle) {
-	ctx := pkcs11.New(lib)
-
-	err := ctx.Initialize()
-	Expect(err).NotTo(HaveOccurred())
-
-	slot := findPKCS11Slot(ctx, label)
-	Expect(slot).Should(BeNumerically(">", 0), "Could not find slot with label %s", label)
-
-	sess, err := ctx.OpenSession(slot, pkcs11.CKF_SERIAL_SESSION|pkcs11.CKF_RW_SESSION)
-	Expect(err).NotTo(HaveOccurred())
-
-	// Login
-	err = ctx.Login(sess, pkcs11.CKU_USER, pin)
-	Expect(err).NotTo(HaveOccurred())
-
-	return ctx, sess
-}
-
-// Identifies pkcs11 slot using specified label
-func findPKCS11Slot(ctx *pkcs11.Ctx, label string) uint {
-	slots, err := ctx.GetSlotList(true)
-	Expect(err).NotTo(HaveOccurred())
-
-	for _, s := range slots {
-		tokInfo, err := ctx.GetTokenInfo(s)
-		Expect(err).NotTo(HaveOccurred())
-
-		if tokInfo.Label == label {
-			return s
-		}
-	}
-
-	return 0
 }
 
 func runQueryInvokeQuery(n *nwo.Network, orderer *nwo.Orderer, peer *nwo.Peer, channel string) {
